Build: tasks moved to files, demo template moved to file
