+++ sources/demo/template.html
@@ -1,0 +1,251 @@
+<!DOCTYPE html>
+<html>
+<head>
+<title>Gridicons</title>
+<meta name="robots" content="noindex">
+<style type="text/css">
+html {
+  background: #fff;
+  color: #767676;
+  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen-Sans", "Ubuntu", "Cantarell", "Helvetica Neue", sans-serif;
+  line-height: 1.15;
+}
+
+a:link, a:visited {
+  color: #999;
+}
+
+a:active {
+  color: #1fc1ad;
+}
+
+h1 {
+  text-align: center;
+  font-size: 24pt;
+}
+
+body > p {
+  text-align: center;
+  margin-bottom: 2em;
+}
+
+body {
+  max-width: 900px;
+  margin: 100px auto;
+}
+
+[type=checkbox] {
+  margin: 0;
+}
+
+.icons {
+  padding: 0 20px;
+  overflow: hidden;
+  margin-bottom: 50px;
+  display: flex;
+  flex-wrap: wrap;
+  justify-content: center;
+}
+
+.icons div {
+  width: 64px;
+  height: 64px;
+  float: left;
+  padding: 6px 2px;
+  position: relative;
+  font-size: 7pt;
+  cursor: pointer;
+  text-align: center;
+}
+
+.icons div p {
+  margin: 0;
+  color: #767676;
+  text-align: center;
+  overflow: hidden;
+  max-height: 2.2em;
+  word-break: break-word;
+}
+
+.icons div p.is-hidden {
+  display: none;
+}
+
+.icons div svg {
+  width: 48px;
+  height: 48px;
+  fill: #000;
+}
+
+.icons div svg.is-small {
+  width: 24px;
+  height: 24px;
+}
+
+.icons div:hover svg {
+  fill: #1fc1ad;
+}
+
+.display-control-group {
+  display: flex;
+  justify-content: space-around;
+  margin-bottom: 20px;
+}
+
+.display-control {
+  display: flex;
+}
+
+.display-control h4 {
+  margin: 0 10px 0 0;
+}
+
+.switch {
+  position: relative;
+  display: inline-block;
+  width: 40px;
+  height: 20px;
+}
+
+.switch input {
+  -webkit-appearance: none;
+  -moz-appearance: none;
+  appearance: none;
+}
+
+.handle {
+  position: absolute;
+  cursor: pointer;
+  top: 0;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  border: 2px solid gray;
+  border-radius: 10px;
+  transition: .4s;
+  box-sizing: border-box;
+}
+
+.handle:before {
+  position: absolute;
+  content: "";
+  height: 12px;
+  width: 12px;
+  left: 2px;
+  bottom: 2px;
+  background: gray;
+  border-radius: 50%;
+  transition: .4s;
+}
+
+input:checked + .handle {
+  border-color: #3AA662;
+}
+
+input:checked + .handle:before {
+  background: #3AA662;
+  transform: translateX(20px);
+}
+
+input:focus + .handle {
+  box-shadow: 0 0 3px #2196F3;
+}
+
+.switch-label {
+  font-size: 10px;
+  font-weight: bold;
+  vertical-align: middle;
+}
+
+.switch-label:after {
+  content: attr(data-off);
+  color: gray;
+}
+.switch-label:before {
+  content: attr(data-on);
+  color: #3AA662;
+  opacity: 0;
+}
+
+input:checked ~ .switch-label:after,
+input:checked ~ .switch-label:before {
+  transition: .4s;
+}
+
+input:checked ~ .switch-label:after {
+  opacity: 0;
+}
+
+input:checked ~ .switch-label:before {
+  opacity: 1;
+}
+</style>
+<script type="text/javascript">
+window.onload = function () {
+    var rows = document.getElementsByTagName( 'svg' );
+    for ( i=0; i<rows.length; i++ ) {
+        rows[i].onclick = function () {
+            var cssClass = this.getAttribute( 'class' );
+            var iconID = cssClass.split(' gridicons-')[1];
+            var suggestion = '<Gridicon icon="'+ iconID +'" />';
+            window.prompt( 'Copy this, paste in your React component.', suggestion );
+        }
+    }
+}
+
+function toggleSize() {
+  var icons = document.querySelectorAll( '.gridicon' );
+  for ( var i = 0; i < icons.length; i++ ) {
+    icons[i].classList.toggle( "is-small" );
+  }
+}
+
+function toggleName() {
+  var icons = document.querySelector( '.icons' ).getElementsByTagName( 'p' );
+  for ( var i = 0; i < icons.length; i++ ) {
+    icons[i].classList.toggle( "is-hidden" );
+  }
+}
+
+// SVG for Everybody
+!function(a,b){"function"==typeof define&&define.amd?define([],function(){return a.svg4everybody=b()}):"object"==typeof module&&module.exports?module.exports=b():a.svg4everybody=b()}(this,function(){function a(a,b,c){if(c){var d=document.createDocumentFragment(),e=!b.hasAttribute("viewBox")&&c.getAttribute("viewBox");e&&b.setAttribute("viewBox",e);for(var f=c.cloneNode(!0);f.childNodes.length;)d.appendChild(f.firstChild);a.appendChild(d)}}function b(b){b.onreadystatechange=function(){if(4===b.readyState){var c=b._cachedDocument;c||(c=b._cachedDocument=document.implementation.createHTMLDocument(""),c.body.innerHTML=b.responseText,b._cachedTarget={}),b._embeds.splice(0).map(function(d){var e=b._cachedTarget[d.id];e||(e=b._cachedTarget[d.id]=c.getElementById(d.id)),a(d.parent,d.svg,e)})}},b.onreadystatechange()}function c(c){function e(){for(var c=0;c<m.length;){var h=m[c],i=h.parentNode,j=d(i);if(j){var n=h.getAttribute("xlink:href")||h.getAttribute("href");if(f&&(!g.validate||g.validate(n,j,h))){i.removeChild(h);var o=n.split("#"),p=o.shift(),q=o.join("#");if(p.length){var r=k[p];r||(r=k[p]=new XMLHttpRequest,r.open("GET",p),r.send(),r._embeds=[]),r._embeds.push({parent:i,svg:j,id:q}),b(r)}else a(i,document.getElementById(q))}}else++c}l(e,67)}var f,g=Object(c),h=/\bTrident\/[567]\b|\bMSIE (?:9|10)\.0\b/,i=/\bAppleWebKit\/(\d+)\b/,j=/\bEdge\/12\.(\d+)\b/;f="polyfill"in g?g.polyfill:h.test(navigator.userAgent)||(navigator.userAgent.match(j)||[])[1]<10547||(navigator.userAgent.match(i)||[])[1]<537;var k={},l=window.requestAnimationFrame||setTimeout,m=document.getElementsByTagName("use");f&&e()}function d(a){for(var b=a;"svg"!==b.nodeName.toLowerCase()&&(b=b.parentNode););return b}return c});svg4everybody();
+</script>
+</head>
+<body>
+
+<h1>Gridicons</h1>
+
+<div class="display-control-group">
+  <div class="display-control">
+    <h4>Small icons</h4>
+    <label class="switch">
+      <input type="checkbox" onchange="toggleSize()">
+      <span class="handle"></span>
+      <span class="switch-label" data-on="On" data-off="Off"></span>
+    </label>
+  </div>
+  <div class="display-control">
+    <h4>Icon names</h4>
+    <label class="switch">
+      <input type="checkbox" onchange="toggleName()" checked>
+      <span class="handle"></span>
+      <span class="switch-label" data-on="On" data-off="Off"></span>
+    </label>
+  </div>
+</div>
+
+<div class="icons">
+{{#each icons}}
+  <div>
+    <svg width="24" height="24" class="gridicon {{name}}">
+    <use xlink:href="gridicons.svg#{{name}}" />
+    </svg>
+    <p>{{title}}</p>
+  </div>
+{{/each}}
+</div>
+
+<p><a href="https://github.com/Automattic/gridicons">GitHub</a></p>
+
+</body>
+</html>
